requirements.dev.txt
bump2version
pre_commit
tox
coverage
django-stubs[compatible-mypy]
